refactor(domain): split Transaction.swift (498→250) (#272)
Extract self-contained neighbours from Transaction.swift into sibling
files along the existing MARK boundaries:
- TransactionType.swift (the enum)
- RecurPeriod.swift (the enum)
- Transaction+Display.swift (Array.uniqued + displayPayee)
- Transaction+Recurrence.swift (nextDueDate, validate, ValidationError)
- RunningBalanceResult.swift (RunningBalanceResult, the conversion-error
  enum, TransactionWithBalance)

The Transaction struct itself, its simple-transaction predicates,
TransactionFilter, and TransactionPage.withRunningBalances stay in
Transaction.swift — those are where the non-file_length baseline entries
are keyed, so moving them would re-fire violations in a new file path.

Drops the file_length baseline entry entirely (498→250). No behaviour
change. No access widenings — every extracted symbol was already
internal. Domain-layer isolation preserved (new files import only
Foundation).

1462 tests in 151 suites pass.
