+++ Domain/Models/Transaction.swift
@@ -4,71 +4,6 @@ import OSLog
 private let transactionLogger = Logger(
   subsystem: "com.moolah.app", category: "Transaction.withRunningBalances")
 
-enum TransactionType: String, Codable, Sendable, CaseIterable {
-  case income
-  case expense
-  case transfer
-  case openingBalance
-
-  /// Whether this transaction type can be manually created or edited by users.
-  /// Opening balance transactions are system-generated and cannot be modified.
-  var isUserEditable: Bool {
-    self != .openingBalance
-  }
-
-  /// Display name for the transaction type
-  var displayName: String {
-    switch self {
-    case .income: return "Income"
-    case .expense: return "Expense"
-    case .transfer: return "Transfer"
-    case .openingBalance: return "Opening Balance"
-    }
-  }
-
-  /// Only types that users can select when creating/editing transactions
-  static var userSelectableTypes: [TransactionType] {
-    [.income, .expense, .transfer]
-  }
-}
-
-enum RecurPeriod: String, Codable, Sendable, CaseIterable {
-  case once = "ONCE"
-  case day = "DAY"
-  case week = "WEEK"
-  case month = "MONTH"
-  case year = "YEAR"
-
-  var displayName: String {
-    switch self {
-    case .once: return "Once"
-    case .day: return "Day"
-    case .week: return "Week"
-    case .month: return "Month"
-    case .year: return "Year"
-    }
-  }
-
-  var pluralDisplayName: String {
-    switch self {
-    case .once: return "Once"
-    case .day: return "Days"
-    case .week: return "Weeks"
-    case .month: return "Months"
-    case .year: return "Years"
-    }
-  }
-}
-
-extension RecurPeriod {
-  /// Human-readable recurrence description, e.g. "Every month" or "Every 2 weeks".
-  func recurrenceDescription(every: Int) -> String {
-    guard self != .once else { return "" }
-    let periodName = every == 1 ? displayName.lowercased() : pluralDisplayName.lowercased()
-    return every == 1 ? "Every \(periodName)" : "Every \(every) \(periodName)"
-  }
-}
-
 struct Transaction: Codable, Sendable, Identifiable, Hashable {
   let id: UUID
   var date: Date
@@ -146,75 +81,6 @@ struct Transaction: Codable, Sendable, Identifiable, Hashable {
     guard let aAcct = a.accountId, let bAcct = b.accountId, aAcct != bAcct else { return false }
     guard b.categoryId == nil && b.earmarkId == nil else { return false }
     return a.instrument != b.instrument
-  }
-}
-
-extension Array where Element: Hashable {
-  /// Returns elements in order of first appearance, removing duplicates.
-  func uniqued() -> [Element] {
-    var seen = Set<Element>()
-    return filter { seen.insert($0).inserted }
-  }
-}
-
-// MARK: - Display Helpers
-
-extension Transaction {
-  /// Builds a display label for the transaction, handling transfers, earmarks, and payees.
-  /// - Parameters:
-  ///   - viewingAccountId: The account the user is viewing from (nil for scheduled/unfiltered views).
-  ///   - accounts: Account lookup collection.
-  ///   - earmarks: Earmark lookup collection.
-  /// - Returns: A human-readable label for the transaction.
-  func displayPayee(
-    viewingAccountId: UUID?, accounts: Accounts, earmarks: Earmarks
-  ) -> String {
-    if isTransfer {
-      if isSimple, let viewingAccountId,
-        let otherLeg = legs.first(where: { $0.accountId != viewingAccountId })
-      {
-        // Account-scoped view: show direction relative to the viewer
-        let otherAccountName =
-          otherLeg.accountId.flatMap { accounts.by(id: $0) }?.name ?? "Unknown Account"
-        let viewingLeg = legs.first(where: { $0.accountId == viewingAccountId })
-        let isOutgoing = (viewingLeg?.quantity ?? 0) < 0
-        let transferLabel =
-          isOutgoing
-          ? "Transfer to \(otherAccountName)"
-          : "Transfer from \(otherAccountName)"
-
-        if let payee, !payee.isEmpty {
-          return "\(payee) (\(transferLabel))"
-        }
-        return transferLabel
-      }
-
-      // No account context (scheduled/upcoming): show "Transfer from A to B"
-      let fromAccount = legs.first(where: { $0.quantity < 0 })?.accountId
-      let toAccount = legs.first(where: { $0.quantity > 0 })?.accountId
-      let fromName = fromAccount.flatMap { accounts.by(id: $0)?.name } ?? "Unknown"
-      let toName = toAccount.flatMap { accounts.by(id: $0)?.name } ?? "Unknown"
-      return "Transfer from \(fromName) to \(toName)"
-    }
-
-    if !isSimple {
-      if let payee, !payee.isEmpty {
-        return "\(payee) (\(legs.count) sub-transactions)"
-      }
-      return "\(legs.count) sub-transactions"
-    }
-
-    if let payee, !payee.isEmpty {
-      return payee
-    }
-
-    let earmarkIds = legs.compactMap(\.earmarkId).uniqued()
-    let earmarkNames = earmarkIds.compactMap { earmarks.by(id: $0)?.name }
-    if !earmarkNames.isEmpty {
-      return "Earmark funds for \(earmarkNames.joined(separator: ", "))"
-    }
-
-    return ""
   }
 }
 
@@ -380,119 +246,5 @@ struct TransactionPage: Sendable {
       rows: result,
       firstConversionError: firstConversionError
     )
-  }
-}
-
-/// The outcome of `TransactionPage.withRunningBalances`: the computed rows plus
-/// the first conversion failure encountered (if any).
-///
-/// Callers that need only the rendered rows can read `rows`. Stores that drive
-/// a user-visible error state should observe `firstConversionError` and publish
-/// it so the user sees a retry path rather than silently blanked balances.
-/// See Rule 11 of `guides/INSTRUMENT_CONVERSION_GUIDE.md`.
-struct RunningBalanceResult: Sendable {
-  let rows: [TransactionWithBalance]
-  let firstConversionError: RunningBalanceConversionError?
-}
-
-/// A typed, Sendable wrapper around the first conversion error encountered
-/// during running-balance computation. Carries the failing transaction id so
-/// diagnostics can correlate logs with the surfaced user-visible error.
-struct RunningBalanceConversionError: LocalizedError, Sendable {
-  let transactionId: UUID
-  let targetInstrumentId: String
-  let underlyingDescription: String
-
-  var errorDescription: String? {
-    "Unable to convert a transaction to \(targetInstrumentId). The running balance is "
-      + "unavailable until the rate source recovers. (\(underlyingDescription))"
-  }
-}
-
-/// A transaction paired with converted leg amounts and the account balance after it was applied.
-///
-/// `displayAmount` and `balance` are `nil` when conversion failed — either for
-/// this transaction's legs or for an earlier transaction in the running-balance
-/// chain. `convertedLegs` is empty in the same case.
-struct TransactionWithBalance: Sendable, Identifiable {
-  let transaction: Transaction
-  let convertedLegs: [ConvertedTransactionLeg]
-  let displayAmount: InstrumentAmount?
-  let balance: InstrumentAmount?
-
-  var id: UUID { transaction.id }
-
-  /// Returns converted legs belonging to the given account.
-  func legs(forAccount accountId: UUID) -> [ConvertedTransactionLeg] {
-    convertedLegs.filter { $0.leg.accountId == accountId }
-  }
-
-  /// Returns converted legs belonging to the given earmark.
-  func legs(forEarmark earmarkId: UUID) -> [ConvertedTransactionLeg] {
-    convertedLegs.filter { $0.leg.earmarkId == earmarkId }
-  }
-}
-
-// MARK: - Recurrence Utilities
-
-extension Transaction {
-  /// Calculates the next due date for a recurring transaction.
-  /// Returns nil if the transaction is not recurring (period is nil or .once).
-  func nextDueDate() -> Date? {
-    guard let period = recurPeriod, let every = recurEvery, period != .once else {
-      return nil
-    }
-
-    let calendar = Calendar.current
-    var components = DateComponents()
-
-    switch period {
-    case .day:
-      components.day = every
-    case .week:
-      components.weekOfYear = every
-    case .month:
-      components.month = every
-    case .year:
-      components.year = every
-    case .once:
-      return nil
-    }
-
-    return calendar.date(byAdding: components, to: date)
-  }
-
-  /// Validates the transaction's fields.
-  func validate() throws {
-    // If either recurPeriod or recurEvery is set, both must be set
-    if (recurPeriod != nil) != (recurEvery != nil) {
-      throw ValidationError.incompleteRecurrence
-    }
-
-    // If recurring, recurEvery must be at least 1
-    if let every = recurEvery, every < 1 {
-      throw ValidationError.invalidRecurEvery
-    }
-
-    if legs.isEmpty {
-      throw ValidationError.noLegs
-    }
-  }
-
-  enum ValidationError: LocalizedError {
-    case incompleteRecurrence
-    case invalidRecurEvery
-    case noLegs
-
-    var errorDescription: String? {
-      switch self {
-      case .incompleteRecurrence:
-        return "Recurrence must have both period and frequency set"
-      case .invalidRecurEvery:
-        return "Recurrence frequency must be at least 1"
-      case .noLegs:
-        return "Transaction must have at least one leg"
-      }
-    }
   }
 }
